+++ pkg/abaputils/abaputils_test.go
@@ -207,3 +207,24 @@ func TestReadServiceKeyAbapEnvironment(t *testing.T) {
 		assert.Error(t, err)
 	})
 }
+
+func TestTimeConverter(t *testing.T) {
+	t.Run("Test example time", func(t *testing.T) {
+		inputDate := "/Date(1585576809000+0000)/"
+		expectedDate := "2020-03-30 14:00:09 +0000 UTC"
+		result := ConvertTime(inputDate)
+		assert.Equal(t, expectedDate, result.String(), "Dates do not match after conversion")
+	})
+	t.Run("Test Unix time", func(t *testing.T) {
+		inputDate := "/Date(0000000000000+0000)/"
+		expectedDate := "1970-01-01 00:00:00 +0000 UTC"
+		result := ConvertTime(inputDate)
+		assert.Equal(t, expectedDate, result.String(), "Dates do not match after conversion")
+	})
+	t.Run("Test unexpected format", func(t *testing.T) {
+		inputDate := "/Date(0012300000001+0000)/"
+		expectedDate := "1970-01-01 00:00:00 +0000 UTC"
+		result := ConvertTime(inputDate)
+		assert.Equal(t, expectedDate, result.String(), "Dates do not match after conversion")
+	})
+}
