+++ pkg/abaputils/manageGitRepositoryUtils_test.go
@@ -1,0 +1,133 @@
+package abaputils
+
+/* func TestPollEntity(t *testing.T) {
+
+	t.Run("Test poll entity: success case", func(t *testing.T) {
+
+		client := &clientMock{
+			BodyList: []string{
+				`{"d" : { "status" : "S" } }`,
+				`{"d" : { "status" : "R" } }`,
+			},
+			Token:      "myToken",
+			StatusCode: 200,
+		}
+		config := abapEnvironmentPullGitRepoOptions{
+			CfAPIEndpoint:     "https://api.endpoint.com",
+			CfOrg:             "testOrg",
+			CfSpace:           "testSpace",
+			CfServiceInstance: "testInstance",
+			CfServiceKeyName:  "testServiceKey",
+			Username:          "testUser",
+			Password:          "testPassword",
+			RepositoryNames:   []string{"testRepo1", "testRepo2"},
+		}
+
+		con := abaputils.ConnectionDetailsHTTP{
+			User:       "MY_USER",
+			Password:   "MY_PW",
+			URL:        "https://api.endpoint.com/Entity/",
+			XCsrfToken: "MY_TOKEN",
+		}
+		status, _ := pollEntity(config.RepositoryNames[0], con, client, 0)
+		assert.Equal(t, "S", status)
+	})
+
+	t.Run("Test poll entity: error case", func(t *testing.T) {
+
+		client := &clientMock{
+			BodyList: []string{
+				`{"d" : { "status" : "E" } }`,
+				`{"d" : { "status" : "R" } }`,
+			},
+			Token:      "myToken",
+			StatusCode: 200,
+		}
+		config := abapEnvironmentPullGitRepoOptions{
+			CfAPIEndpoint:     "https://api.endpoint.com",
+			CfOrg:             "testOrg",
+			CfSpace:           "testSpace",
+			CfServiceInstance: "testInstance",
+			CfServiceKeyName:  "testServiceKey",
+			Username:          "testUser",
+			Password:          "testPassword",
+			RepositoryNames:   []string{"testRepo1", "testRepo2"},
+		}
+
+		con := abaputils.ConnectionDetailsHTTP{
+			User:       "MY_USER",
+			Password:   "MY_PW",
+			URL:        "https://api.endpoint.com/Entity/",
+			XCsrfToken: "MY_TOKEN",
+		}
+		status, _ := pollEntity(config.RepositoryNames[0], con, client, 0)
+		assert.Equal(t, "E", status)
+	})
+
+} */
+
+/* func TestPollEntityCheckoutStep(t *testing.T) {
+
+	t.Run("Test poll entity: success case", func(t *testing.T) {
+
+		client := &clientMock{
+			BodyList: []string{
+				`{"d" : { "status" : "S" } }`,
+				`{"d" : { "status" : "R" } }`,
+			},
+			Token:      "myToken",
+			StatusCode: 200,
+		}
+		config := abapEnvironmentCheckoutBranchOptions{
+			CfAPIEndpoint:     "https://api.endpoint.com",
+			CfOrg:             "testOrg",
+			CfSpace:           "testSpace",
+			CfServiceInstance: "testInstance",
+			CfServiceKeyName:  "testServiceKey",
+			Username:          "testUser",
+			Password:          "testPassword",
+			RepositoryName:    "testRepo1",
+		}
+
+		con := abaputils.ConnectionDetailsHTTP{
+			User:       "MY_USER",
+			Password:   "MY_PW",
+			URL:        "https://api.endpoint.com/Entity/",
+			XCsrfToken: "MY_TOKEN",
+		}
+		status, _ := pollEntity(config.RepositoryName, con, client, 0)
+		assert.Equal(t, "S", status)
+	})
+
+	t.Run("Test poll entity: error case", func(t *testing.T) {
+
+		client := &clientMock{
+			BodyList: []string{
+				`{"d" : { "status" : "E" } }`,
+				`{"d" : { "status" : "R" } }`,
+			},
+			Token:      "myToken",
+			StatusCode: 200,
+		}
+		config := abapEnvironmentCheckoutBranchOptions{
+			CfAPIEndpoint:     "https://api.endpoint.com",
+			CfOrg:             "testOrg",
+			CfSpace:           "testSpace",
+			CfServiceInstance: "testInstance",
+			CfServiceKeyName:  "testServiceKey",
+			Username:          "testUser",
+			Password:          "testPassword",
+			RepositoryName:    "testRepo1",
+		}
+
+		con := abaputils.ConnectionDetailsHTTP{
+			User:       "MY_USER",
+			Password:   "MY_PW",
+			URL:        "https://api.endpoint.com/Entity/",
+			XCsrfToken: "MY_TOKEN",
+		}
+		status, _ := pollEntity(config.RepositoryName, con, client, 0)
+		assert.Equal(t, "E", status)
+	})
+
+} */
